+++ src/Sulu/Bundle/CategoryBundle/Resources/public/js/components/categories/list/main.js
@@ -52,6 +52,8 @@ define(function () {
             this.sandbox.sulu.triggerDeleteSuccessLabel('labels.success.category-delete-desc');
             this.bindCustomEvents();
             this.render();
+
+            this.sandbox.stickyToolbar.enable(this.$el);
         },
 
         bindCustomEvents: function() {
@@ -180,6 +182,10 @@ define(function () {
                     this.sandbox.emit('sulu.labels.success.show', 'labels.success.category-delete-desc', 'labels.success');
                 }.bind(this));
             }.bind(this));
+        },
+
+        destroy: function() {
+            this.sandbox.stickyToolbar.disable(this.$el);
         }
     };
 });