+++ src/Sulu/Bundle/SecurityBundle/Resources/public/js/components/roles/components/list/main.js
@@ -65,6 +65,8 @@ define(function() {
         initialize: function() {
             this.initializeDataGrid();
             bindCustomEvents.call(this);
+
+            this.sandbox.stickyToolbar.enable(this.$el);
         },
 
         initializeDataGrid: function() {
@@ -98,7 +100,10 @@ define(function() {
                 'roles',
                 '#roles-list-info'
             );
+        },
 
+        destroy: function() {
+            this.sandbox.stickyToolbar.disable(this.$el);
         }
     };
 });